check_docs flag

+++ plugin/app/plugins/pipeline_plugin/tools/scm/git/templates/hooks/validators/check_docs.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 function validate() {
   [[ ! -f .git/tempdir/check_docs.sh ]] && return 0
-  .git/tempdir/check_docs.sh  && rm -f .git/tempdir/check_docs
+  if [[ -f .git/tempdir/check_docs ]]; then
+    .git/tempdir/check_docs.sh  && rm -f .git/tempdir/check_docs
+  fi
 }
